Optimize submission queue utilization
Previously the command identifier of a submission queue entry was fixed
and tied directly to the same slot in the callback array. This can lead
to the situation where the submission queue is not fully utilized when
the controller completes requests out of order.

Instead the command identifier of a submission queue entry is now
determined dynamically by searching a free entry in the callback array.
To properly manage this array, the completion and queue head
advancement is moved into the Submission_queue class. Note that it is
very much possible that the submission queue is empty and the callback
array is full. It purely depends on the controller how much commands
can be submitted simultaneously.

Change-Id: I580dd990779a19cb150e7788ff978288b14140c6
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>