Introduce test mode
Some special tests need special handling by the test starter. To
communicate the type of these tests up until now we used the MODE
variable and set this in the .t file. However, since .t files may
invoke the buildsystem and the MODE variable is passed through the
environment this may reconfigure the buildsystem on test execution and
in particular may set a different mode than would have been the default
when compiling parts of the system. To avoid this use a dedicated
TEST_MODE variable to specify that the testing mode should be special.

Currently the only option with any effect is "l4linux". All other
option values are reserved for further use. Not setting the variable
will default to a "normal" test execution.

Change-Id: I34c7b1653deaed4573575b17f5cb28398505705e