Add support for PRP Lists
This change adds support for requests bigger than a page size even if
SGLs are not supported. A compile-time constant can be configured to
specify a number of pages per I/O command that shall be used for
additional PRP entries. The constant can be 0, in which case only the
two PRP entries embedded in the command structure are used (i.e. the old
behavior). A value bigger than 1 will result in a PRP list linked from
multiple pages.

For example, on systems with 4K page size, specifying 1 page for PRP
Lists will allow the guest to use sector-aligned 2M blocks.

Fixes issue #CD-1203.

Change-Id: I42b20d2dd3920eaa41fe06129022356f0667e7cd